+++ core/src/main/java/bisq/core/btc/BaseCurrencyNetwork.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.core.btc.nodes;
+package bisq.core.btc;
 
 import bisq.core.app.BisqEnvironment;
 import bisq.core.provider.fee.FeeService;
