+++ core/src/main/java/bisq/core/btc/BitcoinModule.java
@@ -21,8 +21,7 @@ import bisq.core.app.AppOptionKeys;
 import bisq.core.btc.model.AddressEntryList;
 import bisq.core.btc.model.BalanceModel;
 import bisq.core.btc.nodes.BtcNodes;
-import bisq.core.btc.nodes.RegTestHost;
-import bisq.core.btc.nodes.UserAgent;
+import bisq.core.btc.setup.RegTestHost;
 import bisq.core.btc.setup.WalletsSetup;
 import bisq.core.btc.wallet.BsqCoinSelector;
 import bisq.core.btc.wallet.BsqWalletService;
