+++ core/src/main/java/bisq/core/btc/UserAgent.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.core.btc.nodes;
+package bisq.core.btc;
 
 import javax.inject.Inject;
 import javax.inject.Named;
