+++ desktop/src/test/java/bisq/desktop/util/validation/FiatVolumeValidatorTest.java
@@ -18,7 +18,7 @@
 package bisq.desktop.util.validation;
 
 import bisq.core.app.BisqEnvironment;
-import bisq.core.btc.nodes.BaseCurrencyNetwork;
+import bisq.core.btc.BaseCurrencyNetwork;
 import bisq.core.locale.CurrencyUtil;
 import bisq.core.locale.Res;
